@@ -3,17 +3,8 @@ Hand crafted classes which should undoubtedly be autogenerated from the schema.
 """
 from __future__ import annotations
 
-from typing import (
-    TYPE_CHECKING,
-    Any,
-    Callable,
-    ClassVar,
-    Iterable,
-    Protocol,
-    Type,
-    TypeVar,
-    cast,
-)
+from collections.abc import Callable, Iterable
+from typing import TYPE_CHECKING, Any, ClassVar, Protocol, TypeVar, cast
 
 try:
     from typing import dataclass_transform
@@ -138,7 +129,7 @@ def command(
             validate(instance=instance, schema=response_schema)
             return Response(**instance)
 
-        cls = cast(Type[Command[_R]], cls)
+        cls = cast(type[Command[_R]], cls)
         cls.to_request = to_request
         cls.from_response = from_response
         return frozen(cls)
